# smx-admin/backend/docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/activity-logs": {
      "get": {
        "operationId": "ActivityLogsController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (default: 1)",
            "schema": {
              "minimum": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page (default: 10)",
            "schema": {
              "minimum": 1,
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "adminId",
            "required": false,
            "in": "query",
            "description": "Filter logs by admin ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "entityType",
            "required": false,
            "in": "query",
            "description": "Filter logs by entity type",
            "schema": {
              "enum": [
                "ADMIN",
                "USER",
                "ROLE"
              ],
              "type": "string"
            }
          },
          {
            "name": "activityType",
            "required": false,
            "in": "query",
            "description": "Filter logs by activity type",
            "schema": {
              "enum": [
                "CREATE",
                "UPDATE",
                "DELETE",
                "BLOCK",
                "UNBLOCK",
                "LOGIN",
                "LOGOUT"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of activity logs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedActivityLogsResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - insufficient permissions."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get activity logs with pagination and filters",
        "tags": [
          "Activity Logs"
        ]
      }
    },
    "/activity-logs/export": {
      "get": {
        "operationId": "ActivityLogsController_exportLogs",
        "parameters": [
          {
            "name": "adminId",
            "required": false,
            "in": "query",
            "description": "Filter logs by admin ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "entityType",
            "required": false,
            "in": "query",
            "description": "Filter logs by entity type",
            "schema": {
              "enum": [
                "ADMIN",
                "USER",
                "ROLE"
              ],
              "type": "string"
            }
          },
          {
            "name": "activityType",
            "required": false,
            "in": "query",
            "description": "Filter logs by activity type",
            "schema": {
              "enum": [
                "CREATE",
                "UPDATE",
                "DELETE",
                "BLOCK",
                "UNBLOCK",
                "LOGIN",
                "LOGOUT"
              ],
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "Filter logs from this date (ISO format)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "Filter logs until this date (ISO format)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Activity logs exported successfully",
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - insufficient permissions"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Export activity logs as CSV",
        "tags": [
          "Activity Logs"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "access_token": {
                      "type": "string"
                    },
                    "refresh_token": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          }
        },
        "summary": "Admin login with email and password",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token refreshed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "access_token": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid refresh token"
          }
        },
        "summary": "Refresh access token using refresh token",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successfully logged out"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Logout and revoke all refresh tokens",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/sessions": {
      "get": {
        "operationId": "AuthController_getActiveSessions",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of active sessions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "userAgent": {
                        "type": "string"
                      },
                      "ipAddress": {
                        "type": "string"
                      },
                      "createdAt": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "expiresAt": {
                        "type": "string",
                        "format": "date-time"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all active sessions for the current admin",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/sessions/{tokenId}": {
      "delete": {
        "operationId": "AuthController_revokeSession",
        "parameters": [
          {
            "name": "tokenId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session revoked successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Session not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Revoke a specific session",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "AuthController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns the current admin profile",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "username": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    },
                    "isActive": {
                      "type": "boolean"
                    },
                    "roleName": {
                      "type": "string"
                    },
                    "permissions": {
                      "type": "[string]"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current admin profile",
        "tags": [
          "Authentication"
        ]
      }
    },
...
  },
  "info": {
    "title": "Admin Dashboard API",
    "description": "The Admin Dashboard API documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "AdminUser": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The unique identifier of the admin user",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "username": {
            "type": "string",
            "description": "The username of the admin user",
            "example": "johndoe"
          },
          "email": {
            "type": "string",
            "description": "The email address of the admin user",
            "example": "john.doe@example.com"
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether the admin user is active",
            "example": true
          },
          "role": {
            "description": "The role assigned to the admin user",
            "allOf": [
              {
                "$ref": "#/components/schemas/AdminRole"
              }
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "The date and time when the admin user was created",
            "example": "2024-03-06T21:22:03.123Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "The date and time when the admin user was last updated",
            "example": "2024-03-06T21:22:03.123Z"
          },
          "lastLoginAt": {
            "type": "object",
            "description": "The date and time when the admin user was last logged in",
            "example": "2024-03-06T21:22:03.123Z"
          }
        },
        "required": [
          "id",
          "username",
          "email",
          "isActive",
          "role",
          "createdAt",
          "updatedAt"
        ]
      },
      "AdminRole": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The unique identifier of the role",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "name": {
            "type": "string",
            "description": "The name of the role",
            "example": "SUPER_ADMIN"
          },
          "description": {
            "type": "string",
            "description": "The description of the role",
            "example": "Has full access to all admin features"
          },
          "permissions": {
            "type": "array",
            "description": "The list of permissions assigned to this role",
            "example": [
              "VIEW_ADMINS",
              "CREATE_ADMINS"
            ],
            "items": {
              "type": "string",
              "enum": [
                "VIEW_ADMINS",
                "CREATE_ADMINS",
                "UPDATE_ADMINS",
                "DELETE_ADMINS",
                "BLOCK_ADMINS",
                "VIEW_LOGS",
                "EXPORT_LOGS"
              ]
            }
          },
          "admins": {
            "description": "The list of admins assigned to this role",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminUser"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "The date and time when the role was created",
            "example": "2024-03-06T21:22:03.123Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "The date and time when the role was last updated",
            "example": "2024-03-06T21:22:03.123Z"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "permissions",
          "admins",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "ActivityLog": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The unique identifier of the activity log",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "admin": {
            "description": "The admin user who performed the action",
            "allOf": [
              {
                "$ref": "#/components/schemas/AdminUser"
              }
            ]
          },
          "activityType": {
            "type": "string",
            "description": "The type of activity performed",
            "enum": [
              "CREATE",
              "UPDATE",
              "DELETE",
              "BLOCK",
              "UNBLOCK",
              "LOGIN",
              "LOGOUT"
            ],
            "example": "CREATE"
          },
          "entityType": {
            "type": "string",
            "description": "The type of entity that was affected",
            "enum": [
              "ADMIN",
              "USER",
              "ROLE"
            ],
            "example": "ADMIN"
          },
          "entityId": {
            "type": "string",
            "description": "The ID of the entity that was affected",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "details": {
            "type": "object",
            "description": "Additional details about the activity",
            "example": {
              "username": "johndoe",
              "email": "john.doe@example.com"
            }
          },
          "ipAddress": {
            "type": "string",
            "description": "The IP address from which the activity was performed",
            "example": "192.168.1.1"
          },
          "userAgent": {
            "type": "string",
            "description": "The user agent of the device from which the activity was performed",
            "example": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "The date and time when the activity was logged",
            "example": "2024-03-06T21:22:03.123Z"
          }
        },
        "required": [
          "id",
          "admin",
          "activityType",
          "entityType",
          "entityId",
          "details",
          "ipAddress",
          "userAgent",
          "createdAt"
        ]
      },
      "PaginatedActivityLogsResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ActivityLog"
            }
          },
          "total": {
            "type": "number",
            "example": 100
          },
          "page": {
            "type": "number",
            "example": 1
          },
          "limit": {
            "type": "number",
            "example": 10
          },
          "totalPages": {
            "type": "number",
            "example": 10
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "limit",
          "totalPages"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "The email address of the admin",
            "example": "admin@example.com"
          },
          "password": {
            "type": "string",
            "description": "The password for the admin account",
            "example": "Admin123"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refresh_token": {
            "type": "string",
            "description": "The refresh token obtained during login",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        },
        "required": [
          "refresh_token"
        ]
      },
...
    }
  }
}